orchestrator-v2-migrated_1/requirements.txt
# Interview Orchestrator - Windows 11 / March 2026 Dependencies
flask>=3.1.3
python-dotenv>=1.2.2
google-genai>=1.65.0
pydantic>=2.0